requirements.txt
azure-common
azure-core
azure-identity
azure-mgmt-core
azure-mgmt-resource
msal
mypy
pyinstaller
pytest
requests
ruff
types-requests
types-urllib3
typing_extensions